update settings and degree

+++ backend/app/src/main/java/ch/zhaw/studyflow/domain/curriculum/impls/DegreeManagerImpl.java
@@ -2,21 +2,25 @@ package ch.zhaw.studyflow.domain.curriculum.impls;
 
 import ch.zhaw.studyflow.domain.curriculum.Degree;
 import ch.zhaw.studyflow.domain.curriculum.DegreeManager;
+import ch.zhaw.studyflow.domain.student.StudentManager;
 import ch.zhaw.studyflow.services.persistence.DegreeDao;
 import ch.zhaw.studyflow.services.persistence.SemesterDao;
 import ch.zhaw.studyflow.utils.Validation;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Objects;
 
 public class DegreeManagerImpl implements DegreeManager {
+    private final StudentManager studentManager;
     private final DegreeDao degreeDao;
     private final SemesterDao semesterDao;
 
 
-    public DegreeManagerImpl(DegreeDao degreeDao, SemesterDao semesterDao) {
-        this.degreeDao  = degreeDao;
-        this.semesterDao = semesterDao;
+    public DegreeManagerImpl(StudentManager studentManager, DegreeDao degreeDao, SemesterDao semesterDao) {
+        this.studentManager = studentManager;
+        this.degreeDao      = degreeDao;
+        this.semesterDao    = semesterDao;
     }
 
     @Override
@@ -71,7 +75,25 @@ public class DegreeManagerImpl implements DegreeManager {
         if (degreeId < 0) {
             throw new IllegalArgumentException("Degree ID must be positive");
         }
-        semesterDao.getSemestersForDegree(degreeId).forEach(semester -> semesterDao.deleteSemester(semester.getId()));
-        degreeDao.delete(degreeId);
+
+        Degree degree = degreeDao.read(degreeId);
+        if (degree != null) {
+            semesterDao.getSemestersForDegree(degreeId).forEach(semester -> semesterDao.deleteSemester(semester.getId()));
+            degreeDao.delete(degreeId);
+
+            List<Degree> degreesForStudent = getDegreesForStudent(degree.getOwnerId());
+            final long result = degreesForStudent.stream()
+                    .mapToLong(Degree::getId)
+                    .max()
+                    .orElse(-1L);
+
+            studentManager.getSettings(degree.getOwnerId())
+                    .ifPresent(settings -> {
+                        if (settings.getActiveDegree() == degreeId) {
+                            settings.setActiveDegree(result);
+                            studentManager.updateSettings(settings);
+                        }
+                    });
+        }
     }
 }
